Fixed amount datatype

+++ models.py
@@ -17,8 +17,8 @@ class Employee:
     city: str = None
     state: str = None
     date_of_birth: date = None
-    monthly_salary_base: int = None
-    monthly_salary_left: int = None
+    monthly_salary_base: float = None
+    monthly_salary_left: float = None
     type: str = UserType.EMPLOYEE
     created_at: str = field(default_factory=lambda : datetime.now(ZoneInfo("Asia/Kolkata")).isoformat())
